+++ src/1-one/match/methods/parseMatch/02-parseToken.js
@@ -203,7 +203,12 @@ const parseToken = function (w, opts) {
     //somehow handle encoded-chars?
     w = w.replace('\\*', '*')
     w = w.replace('\\.', '.')
-    obj.word = w.toLowerCase()
+    if (opts.caseSensitive) {
+      obj.use = 'text'
+    } else {
+      w = w.toLowerCase()
+    }
+    obj.word = w
   }
   return obj
 }