+++ changelog.md
@@ -16,6 +16,7 @@ While all _Major_ releases should be reviewed, our only two _large_ releases are
 - **[fix]** - client-side export format for plugins
 - **[new]** - more adjective transformation methods
 - **[new]** - emoji + emoticon tagger
+- **[new]** - case-sensitive match option - `{caseSensitive:true}`
 
 ### 🚨 14.0.0 🚨 &nbsp; [March 2022]
 Major release - see [Release Notes](https://github.com/spencermountain/compromise/releases/tag/14.0.0) for full details